+++ interpreter/src/interpreter-options.spec.js
@@ -4,7 +4,7 @@ import { Interpreter } from './interpreter';
 describe("Interpreter: options", () => {
 
   it('continue flow after selecting an option', () => {
-    const content = parse('\nHey hey\nspeaker: hello\n  * [a]\n   aa\n   ab\n  * [b]\n   ba\n   bb\nend\n');
+    const content = parse('\nHey hey\nspeaker: hello\n  * a\n   aa\n   ab\n  * b\n   ba\n   bb\nend\n');
     const dialogue = Interpreter(content);
 
     expect(dialogue.getContent()).toEqual({ type: 'line', text: 'Hey hey' });
@@ -16,7 +16,7 @@ describe("Interpreter: options", () => {
   });
 
   it('handle sticky and normal options', () => {
-    const content = parse('speaker: hello #name_tag\n  * [a]\n   aa\n   ab\n  * [b $abc #option_tag]\n   ba\n   bb\n  + [c]\n   ca\n   cb\n');
+    const content = parse('speaker: hello #name_tag\n  * a\n   aa\n   ab\n  * b $abc #option_tag\n   ba\n   bb\n  + c\n   ca\n   cb\n');
     const dialogue = Interpreter(content);
 
     expect(dialogue.getContent()).toEqual({ type: 'options', name: 'hello', speaker: 'speaker', tags: [ 'name_tag' ], options: [{ label: 'a' },{ label: 'b', id: 'abc', tags: [ 'option_tag' ] }, { label: 'c' } ] });
@@ -43,7 +43,7 @@ describe("Interpreter: options", () => {
   });
 
   it('use fallback option when others not available', () => {
-    const content = parse('\nHey hey\nspeaker: hello\n  * a\n  > b\nend\n');
+    const content = parse('\nHey hey\nspeaker: hello\n  *= a\n  >= b\nend\n');
     const dialogue = Interpreter(content);
 
     expect(dialogue.getContent()).toEqual({ type: 'line', text: 'Hey hey' });
@@ -58,7 +58,7 @@ describe("Interpreter: options", () => {
   });
 
   it('expose special variable OPTIONS_COUNT', () => {
-    const content = parse('speaker: hello\n  * [a]\n   a %OPTIONS_COUNT%\n  * [b]\n   b %OPTIONS_COUNT%\n  * [c %OPTIONS_COUNT% left]\n   c %OPTIONS_COUNT%\n');
+    const content = parse('speaker: hello\n  * a\n   a %OPTIONS_COUNT%\n  * b\n   b %OPTIONS_COUNT%\n  * c %OPTIONS_COUNT% left\n   c %OPTIONS_COUNT%\n');
     const dialogue = Interpreter(content);
 
     expect(dialogue.getContent()).toEqual({ type: 'options', name: 'hello', speaker: 'speaker', options: [{ label: 'a' },{ label: 'b' }, { label: 'c 3 left' } ] });
@@ -87,13 +87,13 @@ describe("Interpreter: options", () => {
   it('use special variable OPTIONS_COUNT as condition', () => {
     const content = parse(`
 hello %OPTIONS_COUNT%
-    * [Yes]
+    * Yes
       yep
       <-
-    * [No]
+    * No
       nope
       <-
-    + { OPTIONS_COUNT > 1 } [What?]
+    + { OPTIONS_COUNT > 1 } What?
       wat
       <-
 `);
@@ -121,7 +121,7 @@ hello %OPTIONS_COUNT%
   });
 
   it('fails when selecting wrong index', () => {
-    const content = parse('hello $123\n * [a]\n  aa\n * [b]\n  ba\n');
+    const content = parse('hello $123\n * a\n  aa\n * b\n  ba\n');
     const dialogue = Interpreter(content);
     expect(dialogue.getContent()).toEqual({ id: '123', type: 'options', name: 'hello', options: [{ label: 'a' }, { label: 'b' } ] });
     expect(() => dialogue.choose(66)).toThrow(/Index 66 not available./);
@@ -132,7 +132,7 @@ hello %OPTIONS_COUNT%
     const content = parse(`
 { set europeTopicsTalked = 0 }
 Vincent: What do you want to know?
-  * [Is hash legal there?]
+  * Is hash legal there?
     Jules: is hash legal there?
     Vincent: yes, but is ain't a hundred percent legal.
              I mean you can't walk into a restaurant, roll a joint,
@@ -147,19 +147,19 @@ Vincent: What do you want to know?
              Searching you is a right that the cops in Amsterdam don't have."
     Jules: That did it, man - I'm f**n' goin', that's all there is to it.
     <-
-  + { europeTopicsTalked < 4 } [Something about Europe.]
+  + { europeTopicsTalked < 4 } Something about Europe.
     Vincent: You know what the funniest thing about Europe is?
     Jules: what?
     Vincent: It's the little differences. A lotta the same sh*t we got here, they
              they got there, but there they're a little different.
     Jules: Examples?
-      * [You can buy a beer in a movie theatre.]
+      * You can buy a beer in a movie theatre.
         Vincent: Well, in Amsterdam, you can buy beer in a
                  movie theatre. And I don't mean in a paper
                  cup either. They give you a glass of beer,
         { set europeTopicsTalked += 1}
         <-
-      * [You know what they call a Quarter Pounder with Cheese in Paris?]
+      * You know what they call a Quarter Pounder with Cheese in Paris?
         Vincent: You know what they call a Quarter Pounder with Cheese in Paris?
         Jules: They don't call it a Quarter Pounder with Cheese?
         Vincent: No, they got the metric system there, they wouldn't know what
@@ -171,11 +171,11 @@ Vincent: What do you want to know?
         { set quarterPounderTalkCompleted = true }
         { set europeTopicsTalked += 1}
         <-
-      * { quarterPounderTalkCompleted } [What do they call a Whopper?]
+      * { quarterPounderTalkCompleted } What do they call a Whopper?
         Jules: What do they call a Whopper?
         Vincent: I dunno, I didn't go into a Burger King.
         { set europeTopicsTalked += 1}
-      * [What they put on the french fries instead of ketchup.]
+      * What they put on the french fries instead of ketchup.
         Vincent: You know what they put on french fries in Holland
                  instead of ketchup?
         Jules: What?
@@ -186,11 +186,11 @@ Vincent: What do you want to know?
         Jules: Uuccch!
         { set europeTopicsTalked += 1}
         <-
-      + { OPTIONS_COUNT > 1 } [I'm suddenly not interested anymore.]
+      + { OPTIONS_COUNT > 1 } I'm suddenly not interested anymore.
         Jules: We talk about this another time.
     { set europeTalkCompleted = true }
     <-
-  + { OPTIONS_COUNT > 1 } [Nah, maybe another time]
+  + { OPTIONS_COUNT > 1 } Nah, maybe another time
         (
           - Vincent: Alright!
           - Vincent: No problem!
@@ -227,8 +227,8 @@ Jules: Let's get to work!
     const content = parse(`
 Hey hey
 speaker: hello
-      * a { set something = true }
-      * b { when not something }
+      *= a { set something = true }
+      *= b { when not something }
 hey
 `);
     const dialogue = Interpreter(content);
@@ -246,9 +246,9 @@ hey
     const content = parse(`
 Hey hey
 speaker: hello
-      * a { set something = true }
+      *= a { set something = true }
         hey you
-      * b
+      *= b
 hey
 `);
     const dialogue = Interpreter(content);
@@ -265,9 +265,9 @@ hey
     const content = parse(`
 Hey hey
 speaker: hello
-      * { not something } a { set something = true }
-      * b { when not something }
-      * { set something = 2 } c  { when a == 42 }
+      *= { not something } a { set something = true }
+      *= b { when not something }
+      *= { set something = 2 } c  { when a == 42 }
 hey
 `);
     const dialogue = Interpreter(content);