+++ parser/src/parser-logic.spec.js
@@ -443,9 +443,9 @@ describe('parse: logic', () => {
 
     it('conditional option', () => {
       const result = parse(`
-* { some_var } option 1
-* option 2 { when some_var }
-* { some_other_var } option 3
+*= { some_var } option 1
+*= option 2 { when some_var }
+*= { some_other_var } option 3
 `);
       const expected = createDocPayload([{
         type: 'options',
@@ -824,9 +824,9 @@ describe('parse: logic', () => {
 
     it('options assignment', () => {
       const result = parse(`
-* { set a = 2 } option 1
-* option 2 { set b = 3 }
-* { set c = 4 } option 3
+*= { set a = 2 } option 1
+*= option 2 { set b = 3 }
+*= { set c = 4 } option 3
 `);
       const expected = createDocPayload([{
         type: 'options',
@@ -968,9 +968,9 @@ describe('parse: logic', () => {
 
     it('options trigger', () => {
       const result = parse(`
-* { trigger a } option 1
-* option 2 { trigger b }
-* { trigger c } option 3
+*= { trigger a } option 1
+*= option 2 { trigger b }
+*= { trigger c } option 3
 `);
       const expected = createDocPayload([{
         type: 'options',