+++ cli/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## 2.0.0 (2021-11-21)
+
+### Breaking changes
+
+Options do not display first line by default. Square brackets are not allowed anymore.
+
+### Changed
+
+- Updated @clyde-lang/parser to v2.0.0 and @clyde-lang/interpreter to v2.0.0.
+- Option lines are not included as content anymore.
+- Supports `=` to keep old behaviour of showing the main line.
+- Brackets (`[]`) for display only options are not required or allowed.
+
 ## 1.0.1 (2021-10-21)
 
 ### Breaking changes